requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
Flask-Migrate>=4.0

# Database
SQLAlchemy==2.0.23

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Security
Werkzeug==2.3.7
itsdangerous==2.1.2

# Development Tools
python-dotenv==1.0.0

# Utilities
Jinja2==3.1.2
MarkupSafe==2.1.3
WTForms==3.1.1


Flask-Dance[sqla]>=7.0 

# Environment Variables

python-dotenv>=1.0




